fix: Amend PR comments
